Prints entire elements of list_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t lst_node = 0;
+	unsigned int lst_node = 0;
 
 	while (h != NULL)
 	{
